pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "llm-finetune"
version = "0.1.0"
description = "Parameter-efficient LLM fine-tuning with LoRA/QLoRA and quantised inference"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["llm", "fine-tuning", "lora", "qlora", "inference", "quantisation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "datasets>=2.18.0",
    "tokenizers>=0.19.0",
    "accelerate>=0.29.0",
    "safetensors>=0.4.3",
    "huggingface-hub>=0.22.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
]

[project.optional-dependencies]
train = [
    "bitsandbytes>=0.43.0",
    "trl>=0.8.0",
    "wandb>=0.16.0",
    "tensorboard>=2.16.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "nltk>=3.8.1",
]
flash-attn = [
    "flash-attn>=2.5.0",
]
inference = [
    "llama-cpp-python>=0.2.60",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.9.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
all = [
    "llm-finetune[train,flash-attn,inference,dev]",
]

[project.scripts]
llm-train   = "scripts.train:app"
llm-eval    = "scripts.evaluate:app"
llm-quant   = "scripts.quantise:app"
llm-merge   = "scripts.merge_adapter:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
